+++ db/sql/access_key.go
@@ -11,6 +11,9 @@ import (
 
 func (d *SqlDb) GetAccessKey(projectID int, accessKeyID int) (key db.AccessKey, err error) {
 	err = d.getObject(projectID, db.AccessKeyProps, accessKeyID, &key)
+	if err == nil {
+		key.Empty = key.IsEmpty()
+	}
 	return
 }